+++ src/main/java/com/linkall/vance/core/Sink.java
@@ -1,4 +1,4 @@
-package com.linkall.core;
+package com.linkall.vance.core;
 
 public interface Sink {
     void start() throws Exception;
