+++ src/main/java/com/linkall/vance/core/http/HttpServer.java
@@ -14,10 +14,10 @@
  * limitations under the License.
  *
  */
-package com.linkall.core.http;
+package com.linkall.vance.core.http;
 
-import com.linkall.core.Adapter;
-import com.linkall.core.Adapter2;
+import com.linkall.vance.core.Adapter;
+import com.linkall.vance.core.Adapter2;
 import io.cloudevents.CloudEvent;
 import io.vertx.core.Handler;
 import io.vertx.core.buffer.Buffer;
@@ -43,7 +43,7 @@ public interface HttpServer {
      * {@link HttpServer#simpleHandler(Adapter2)}
      * to simply tell SDK how to generate a CloudEvent caring about nothing else.
      * @param handler
-     * @param <T> extends {@link io.vertx.core.Handler} and {@link com.linkall.core.Adapter}
+     * @param <T> extends {@link io.vertx.core.Handler} and {@link Adapter}
      */
     <T extends Handler<RoutingContext> & Adapter>void handler(T handler);
     //route to a specific path
